requirements.txt
python-telegram-bot[webhooks]
pycryptodomex
requests
python-dotenv
beautifulsoup4
